# Store NVFP4 block scales in swwizzled layout on tensor