--- a/torchao/prototype/mx_formats/nvfp4_tensor.py
+++ b/torchao/prototype/mx_formats/nvfp4_tensor.py
@@ -4,6 +4,7 @@
 # This source code is licensed under the BSD 3-Clause license found in the
 # LICENSE file in the root directory of this source tree.
 
+import sys
 from enum import Enum
 from typing import Any, Callable, Dict, Optional
 
@@ -21,8 +22,8 @@ from torchao.prototype.mx_formats.mx_tensor import (
     tensor_size_fp4x2_to_hp,
     tensor_size_hp_to_fp4x2,
 )
-from torchao.prototype.mx_formats.utils import to_blocked
-from torchao.utils import fill_defaults
+from torchao.prototype.mx_formats.utils import from_blocked, to_blocked
+from torchao.utils import ceil_div, fill_defaults
 
 E4M3_EPS = torch.finfo(torch.float8_e4m3fn).tiny
 
@@ -54,11 +55,12 @@ class NVFP4Tensor(torch.Tensor):
     quantization algorithm for FP4 data with UE4M3 scales.
 
     Attributes:
-        _scale_e4m3: Blockwise scales in float8_e4m3fn format
+        _scale_e4m3: Blockwise scales in float8_e4m3fn format (may be swizzled)
         _per_tensor_scale: Optional global per-tensor scale in float32 format
         _data: Packed FP4 data (2 values per byte)
         _block_size: Block size for quantization (fixed at 16)
         _orig_dtype: Original tensor dtype before quantization
+        _is_swizzled_scales: Whether scales are stored in swizzled (blocked) format
         mm_config: Matrix multiplication configuration
     """
 
@@ -67,6 +69,7 @@ class NVFP4Tensor(torch.Tensor):
     _data: torch.Tensor
     _block_size: int
     _orig_dtype: torch.dtype
+    _is_swizzled_scales: bool
     mm_config: NVFP4MMConfig
 
     def __new__(
@@ -77,12 +80,14 @@ class NVFP4Tensor(torch.Tensor):
         block_size,
         orig_dtype,
         mm_config=NVFP4MMConfig.DYNAMIC,
+        is_swizzled_scales=False,
     ):
-        # FP4 tensor size handling
+        # FP4 tensor size handling two paths, contiguous or not
         new_size = data_bits.size()
+
         new_size = tensor_size_fp4x2_to_hp(
             new_size,
-            data_bits.is_contiguous(),
+            data_bits.stride(0) > data_bits.stride(1),
         )
 
         self = torch.Tensor._make_wrapper_subclass(
@@ -94,6 +99,7 @@ class NVFP4Tensor(torch.Tensor):
         )
 
         self._scale_e4m3 = blockwise_scales
+        self._is_swizzled_scales = is_swizzled_scales
         self._per_tensor_scale = per_tensor_scale
         self._data = data_bits
         self._block_size = block_size
@@ -118,14 +124,17 @@ class NVFP4Tensor(torch.Tensor):
         block_size: int = 16,
         per_tensor_scale: Optional[torch.Tensor] = None,
         mm_config: NVFP4MMConfig = NVFP4MMConfig.DYNAMIC,
+        is_swizzled_scales: bool = False,
     ):
         """Convert high precision tensor to NVFP4 format.
 
         Args:
             data_hp: High precision input tensor (bfloat16 or float32)
             block_size: Block size for quantization (must be 16)
-            per_tensor_amax: Optional pre-computed absolute maximum for calibration.
+            per_tensor_scale: Optional pre-computed absolute maximum for calibration.
                 If provided, uses per-tensor scaling. If None, uses block-wise scaling only.
+            mm_config: Matrix multiplication configuration
+            is_swizzled_scales: If True, store scales in swizzled format for faster matrix multiplication
 
         Returns:
             NVFP4Tensor: Quantized tensor in NVFP4 format
@@ -133,6 +142,12 @@ class NVFP4Tensor(torch.Tensor):
         blockwise_scales, data_lp = nvfp4_quantize(
             data_hp, block_size, per_tensor_scale
         )
+
+        if is_swizzled_scales:
+            M, K = data_hp.shape[0], data_hp.shape[1]
+            scale_shape = (M, K // block_size)
+            blockwise_scales = to_blocked(blockwise_scales.view(scale_shape)).flatten()
+
         return NVFP4Tensor(
             blockwise_scales,
             per_tensor_scale,
@@ -140,12 +155,14 @@ class NVFP4Tensor(torch.Tensor):
             block_size,
             data_hp.dtype,
             mm_config,
+            is_swizzled_scales,
         )
 
     def __tensor_flatten__(self):
         ctx = {
             "_block_size": self._block_size,
             "_orig_dtype": self._orig_dtype,
+            "_is_swizzled_scales": self._is_swizzled_scales,
             "mm_config": self.mm_config,
         }
         tensor_list = ["_scale_e4m3", "_data"]
@@ -182,6 +199,7 @@ class NVFP4Tensor(torch.Tensor):
             metadata["_block_size"],
             metadata["_orig_dtype"],
             metadata["mm_config"],
+            metadata.get("_is_swizzled_scales", False),
         )
 
     # Do not force the NVFP4Tensor type on the returned tensor
@@ -196,7 +214,7 @@ class NVFP4Tensor(torch.Tensor):
         Returns:
             torch.Tensor: Dequantized tensor in the target dtype
         """
-        is_transposed = not self._data.is_contiguous()
+        is_transposed = self._data.stride(0) < self._data.stride(1)
         if is_transposed:
             M, K = self.shape[1], self.shape[0]
         else:
@@ -221,10 +239,21 @@ class NVFP4Tensor(torch.Tensor):
         Returns:
             torch.Tensor: Scales of the NVFP4Tensor
         """
+        is_transposed = self._data.stride(0) < self._data.stride(1)
+        if is_transposed:
+            M, K = self.shape[1], self.shape[0]
+        else:
+            M, K = self.shape[0], self.shape[1]
+
+        if self._is_swizzled_scales:
+            scale_e4m3 = from_blocked(self._scale_e4m3, M, K // self._block_size)
+        else:
+            scale_e4m3 = self._scale_e4m3
+
         return (
-            self._scale_e4m3.to(self._orig_dtype)
+            scale_e4m3.to(self._orig_dtype)
             if not self._per_tensor_scale
-            else self._per_tensor_scale * self._scale_e4m3.to(self._orig_dtype)
+            else self._per_tensor_scale * scale_e4m3.to(self._orig_dtype)
         )
 
     @classmethod
@@ -238,7 +267,6 @@ class NVFP4Tensor(torch.Tensor):
         Returns:
             bool: True if both tensors have identical metadata, False otherwise
         """
-        # Check per_tensor_scale equality
         per_tensor_scale_equal = (
             self._per_tensor_scale is None and src._per_tensor_scale is None
         ) or (self._per_tensor_scale.shape == src._per_tensor_scale.shape)
@@ -248,6 +276,7 @@ class NVFP4Tensor(torch.Tensor):
             and isinstance(src, NVFP4Tensor)
             and self._block_size == src._block_size
             and self._orig_dtype == src._orig_dtype
+            and self._is_swizzled_scales == src._is_swizzled_scales
             and self._scale_e4m3.shape == src._scale_e4m3.shape
             and per_tensor_scale_equal
             and self._data.shape == src._data.shape
@@ -292,6 +321,7 @@ def nvfp4_to_copy(func, types, args, kwargs):
             tensor._block_size,
             dtype,
             tensor.mm_config,
+            tensor._is_swizzled_scales,
         )
         return res
 
@@ -335,45 +365,165 @@ def nvfp4_slice(func, types, args, kwargs):
     assert x._data.is_contiguous(), "Only support contiguous data for now"
 
     M, K = x.shape[0], x.shape[1]
-    scale_shaped = x._scale_e4m3.view(M, K // x._block_size)
 
-    if dim == 0:
-        # Slicing along the first dimension (rows)
-        sliced_scale = aten.slice.Tensor(scale_shaped, dim, start, end, step).flatten()
-        sliced_data = aten.slice.Tensor(x._data, dim, start, end, step)
-    elif dim == 1:
-        # Slicing along reduction dim - must align with block boundaries
-        if start is not None:
-            assert start % x._block_size == 0, (
-                f"Start index {start} must be a multiple of block_size {x._block_size}"
+    if x._is_swizzled_scales:
+        scale_rows = M
+        scale_cols = K // x._block_size
+        n_row_blocks = ceil_div(scale_rows, 128)
+        n_col_blocks = ceil_div(scale_cols, 4)
+        elements_per_block = 32 * 16  # 512 elements
+
+        if dim == 0:
+            # Row slicing
+            # Handle sys.maxsize (default slice end)
+            if end == sys.maxsize:
+                end = M
+
+            # Check if start/end align with 128-row boundaries
+            if start is not None and start % 128 != 0:
+                raise RuntimeError(
+                    f"Row slicing of NVFP4Tensor with swizzled scales requires "
+                    f"start index to be a multiple of 128, got {start}"
+                )
+            if end is not None and end != M and end % 128 != 0:
+                raise RuntimeError(
+                    f"Row slicing of NVFP4Tensor with swizzled scales requires "
+                    f"end index to be a multiple of 128 or equal to tensor size {M}, got {end}"
+                )
+
+            # Calculate which row blocks to keep
+            start_block = 0 if start is None else start // 128
+            end_block = n_row_blocks if end is None or end >= M else end // 128
+
+            # The swizzled tensor has shape (n_row_blocks * n_col_blocks * 32 * 16,)
+            blocks_per_row = n_col_blocks
+            start_idx = start_block * blocks_per_row * elements_per_block
+            end_idx = (
+                end_block * blocks_per_row * elements_per_block
+                if end_block < n_row_blocks
+                else None
             )
 
-        if end is not None:
-            assert end % x._block_size == 0, (
-                f"End index {end} must be a multiple of block_size {x._block_size}"
+            sliced_scale = aten.slice.Tensor(x._scale_e4m3, 0, start_idx, end_idx, 1)
+            sliced_data = aten.slice.Tensor(x._data, 0, start, end, step)
+
+        elif dim == 1:
+            # Column slicing
+            # Handle sys.maxsize (default slice end)
+            if end == sys.maxsize:
+                end = K
+
+            # Check if start/end align with 64-column boundaries (4 scale columns * 16 block_size)
+            if start is not None and start % 64 != 0:
+                raise RuntimeError(
+                    f"Column slicing of NVFP4Tensor with swizzled scales requires "
+                    f"start index to be a multiple of 64, got {start}"
+                )
+            if end is not None and end != K and end % 64 != 0:
+                raise RuntimeError(
+                    f"Column slicing of NVFP4Tensor with swizzled scales requires "
+                    f"end index to be a multiple of 64 or equal to tensor size {K}, got {end}"
+                )
+
+            # Also check FP4 packing alignment
+            if start is not None and start % 2 != 0:
+                raise RuntimeError(f"Start index {start} must be even for FP4 packing")
+            if end is not None and end != K and end % 2 != 0:
+                raise RuntimeError(f"End index {end} must be even for FP4 packing")
+
+            # Calculate which column blocks to keep
+            start_scale_col = 0 if start is None else start // 16
+            end_scale_col = scale_cols if end is None or end >= K else end // 16
+
+            start_col_block = start_scale_col // 4
+            end_col_block = end_scale_col // 4
+
+            # Verify the end aligns with block boundary
+            if end_scale_col % 4 != 0:
+                raise RuntimeError(
+                    f"Column slicing end index {end} does not align with scale block boundaries. "
+                    f"End must result in a multiple of 4 scale columns (64 data columns)."
+                )
+
+            if start_col_block == 0 and end_col_block == n_col_blocks:
+                # Full width - no slicing needed
+                sliced_scale = x._scale_e4m3
+            else:
+                # Extract specific column blocks from each row block
+                # Each row block in swizzled format contains n_col_blocks chunks of (32, 16)
+                elements_per_row_block = n_col_blocks * elements_per_block
+
+                # Build list of slices to extract
+                slices_to_extract = []
+                for row_block in range(n_row_blocks):
+                    row_start = row_block * elements_per_row_block
+                    col_start = row_start + start_col_block * elements_per_block
+                    col_end = row_start + end_col_block * elements_per_block
+                    slices_to_extract.append(x._scale_e4m3[col_start:col_end])
+
+                # Concatenate all the slices
+                sliced_scale = torch.cat(slices_to_extract, dim=0)
+
+            # Slice the data tensor
+            packed_start = None if start is None else start // 2
+            packed_end = None if end is None else end // 2
+            sliced_data = aten.slice.Tensor(
+                x._data, dim, packed_start, packed_end, step
             )
 
-        sliced_data = aten.slice.Tensor(x._data, dim, start, end, step)
+        else:
+            raise ValueError(
+                f"NVFP4Tensor only supports slicing along dimensions 0 and 1, got dim={dim}"
+            )
 
-        # Calculate which scale blocks to keep
-        start_block = 0 if start is None else start // x._block_size
-        end_block = None if end is None else end // x._block_size
-
-        # Slice the scale tensor accordingly
-        sliced_scale = aten.slice.Tensor(scale_shaped, 1, start_block, end_block, step)
     else:
-        raise ValueError(
-            f"NVFP4Tensor only supports slicing along dimensions 0 and 1, got dim={dim}"
-        )
+        scale_shaped = x._scale_e4m3.view(M, K // x._block_size)
 
-    return NVFP4Tensor(
+        if dim == 0:
+            sliced_scale = aten.slice.Tensor(scale_shaped, dim, start, end, step)
+            sliced_data = aten.slice.Tensor(x._data, dim, start, end, step)
+
+        elif dim == 1:
+            if start is not None:
+                assert start % x._block_size == 0, (
+                    f"Start index {start} must be a multiple of block_size {x._block_size}"
+                )
+                assert start % 2 == 0, (
+                    f"Start index {start} must be even for FP4 packing"
+                )
+
+            if end is not None and end != sys.maxsize:
+                assert end % x._block_size == 0, (
+                    f"End index {end} must be a multiple of block_size {x._block_size}"
+                )
+                assert end % 2 == 0, f"End index {end} must be even for FP4 packing"
+
+            packed_start = None if start is None else start // 2
+            packed_end = None if end is None else end // 2
+            sliced_data = aten.slice.Tensor(
+                x._data, dim, packed_start, packed_end, step
+            )
+
+            start_block = 0 if start is None else start // x._block_size
+            end_block = None if end is None else end // x._block_size
+            sliced_scale = aten.slice.Tensor(
+                scale_shaped, 1, start_block, end_block, step
+            )
+
+        sliced_scale = sliced_scale.flatten()
+
+    # Create result tensor
+    result = NVFP4Tensor(
         sliced_scale,
-        x._per_tensor_scale,  # Unchanged per-tensor scale
+        x._per_tensor_scale,
         sliced_data,
         x._block_size,
         x._orig_dtype,
         x.mm_config,
+        x._is_swizzled_scales,
     )
+
+    return return_and_correct_aliasing(func, args, kwargs, result)
 
 
 @implements([aten.t.default])
@@ -387,6 +537,7 @@ def nvfp4_t(func, types, args, kwargs):
         old._block_size,
         old._orig_dtype,
         old.mm_config,
+        old._is_swizzled_scales,
     )
     return new
 
@@ -404,6 +555,7 @@ def nvfp4_view_op(func, types, args, kwargs):
         args[0]._block_size,
         args[0]._orig_dtype,
         args[0].mm_config,
+        args[0]._is_swizzled_scales,
     )
 
 
@@ -423,10 +575,17 @@ def _addmm_nvfp4_dispatch(
     N = b.shape[1]
 
     # Swizzle Dizzle
-    a_scale = a._scale_e4m3.view(M, K // a._block_size)
-    b_scale = b._scale_e4m3.view(N, K // b._block_size)
-    a_scale_blocked = to_blocked(a_scale)
-    b_scale_blocked = to_blocked(b_scale)
+    if a._is_swizzled_scales:
+        a_scale_blocked = a._scale_e4m3  # Already swizzled
+    else:
+        a_scale = a._scale_e4m3.view(M, K // a._block_size)
+        a_scale_blocked = to_blocked(a_scale)
+
+    if b._is_swizzled_scales:
+        b_scale_blocked = b._scale_e4m3  # Already swizzled
+    else:
+        b_scale = b._scale_e4m3.view(N, K // b._block_size)
+        b_scale_blocked = to_blocked(b_scale)
 
     # Merge double quant scales into 1 scale for Scale_In^D
     if a._per_tensor_scale is not None:
@@ -571,8 +730,8 @@ def nvfp4_quantize(
     assert data_hp.dtype in (torch.bfloat16, torch.float), (
         f"{data_hp.dtype} not supported"
     )
-    assert data_hp.numel() % block_size == 0, "unsupported"
-    assert data_hp.is_contiguous(), "unsupported"
+    assert data_hp.size(-1) % block_size == 0, "K dim must be divisible by block_size"
+    assert data_hp.is_contiguous(), "Only support contiguous data for now"
     assert block_size == 16, "NVFP4 requires block_size=16"
 
     orig_shape = data_hp.shape