--- a/torchao/utils.py
+++ b/torchao/utils.py
@@ -698,6 +698,10 @@ def check_xpu_version(device, version="2.8.0"):
     return device == "xpu" and compare_versions(torch.__version__, version) >= 0
 
 
+def ceil_div(a, b):
+    return (a + b - 1) // b
+
+
 TORCH_VERSION_AFTER_2_5 = _torch_version_at_least("2.5.0.dev")
 TORCH_VERSION_AFTER_2_4 = _torch_version_at_least("2.4.0.dev")
 TORCH_VERSION_AFTER_2_3 = _torch_version_at_least("2.3.0.dev")