Stacked PRs:
 * #2439
 * __->__#2438


--- --- ---

Store NVFP4 block scales in swizzled layout on tensor

For llama3 70b no TP sizes w/ 1024 tokens: 15% E2E speedup

## In eager Before: https://fburl.com/7w3j6b1q
nvfp4 
<img width="693" alt="Screenshot 2025-06-24 at 4 02 42 PM" src="https://github.com/user-attachments/assets/2c4c5b94-3162-4733-8c97-6f40f47b14aa" />
Runtime: 2436.98 μs per iteration

## In eager After: https://fburl.com/s7ggvm94
nvfp4 Runtime: 2356.77 μs per iteration
<img width="349" alt="Screenshot 2025-06-24 at 4 03 27 PM" src="https://github.com/user-attachments/assets/f08d3c6c-efe2-4381-9573-a53d33b6f5e4" />


# In compile


## Before:  https://fburl.com/1gvfjjlu
nvfp4 Runtime: 576.14 μs per iteration
<img width="568" alt="Screenshot 2025-06-24 at 4 11 36 PM" src="https://github.com/user-attachments/assets/6c4066cb-a1f4-45e6-867c-70bc91d93f80" />


## After: https://fburl.com/usp1xelj
nvfp4 Runtime: 486.69 μs per iteration
<img width="668" alt="Screenshot 2025-06-24 at 4 11 55 PM" src="https://github.com/user-attachments/assets/02deebee-8087-42b5-a926-8c88273ad571" />


```
Throughput: 47.12 requests/s, 19998.00 total tokens/s, 9635.87 output tokens/s
Total num prompt tokens:  225190
Total num output tokens:  209407
```