@drisspg 2025-06-26
Yup this is for inference

```Py

import torch
from typing import Literal

from transformer_nuggets.misc.mlp import FeedForward
from transformer_nuggets.utils.benchmark import (
    profiler,
    benchmark_do_bench_in_microseconds,
    benchmark_cuda_function_in_microseconds,
)

from torchao.quantization import quantize_
from torchao.prototype.mx_formats.mx_subclass import (
    MXFPInferenceConfig,
    NVFP4InferenceConfig,
)
from torchao.prototype.mx_formats import MXGemmKernelChoice
import os
from rich import print

from jsonargparse import CLI

torch._logging.set_logs(fusion=True)
torch._logging.set_logs(graph=True)
os.environ["TORCHINDUCTOR_FORCE_DISABLE_CACHES"] = "1"


def main(
    num_tokens: int = 1024,
    flavor: str = "8B",
    mode: str = "mxfp8",
    bf16: bool = False,
    setting: Literal["max-autotune", "inductor"] = "max-autotune",
    backend: Literal["inductor", "eager"] = "inductor",
    use_triton: bool = False,
):
    device = "cuda"
    dtype = torch.bfloat16
    assert flavor in ("8B", "70B", "405B")
    assert mode in ("mxfp8", "mxfp4", "nvfp4")
    if bf16:
        model = FeedForward.llama3_mlp(flavor).to(device=device, dtype=dtype)
        inpt = model.get_input(num_tokens, device, dtype)
        model = torch.compile(
            model, fullgraph=True, dynamic=False, mode=setting, backend=backend
        )

        for _ in range(5):
            model(inpt)
        with profiler("data/bf16", with_stack=True) as p:
            for _ in range(5):
                model(inpt)
                p.step()
        runtime_us = benchmark_cuda_function_in_microseconds(lambda: model(inpt))
        print(
            f"[bold green]Bf16 Runtime:[/bold green] {runtime_us:.2f} μs per iteration"
        )

    model = FeedForward.llama3_mlp(flavor).to(device=device, dtype=dtype)
    inpt = model.get_input(num_tokens, device, dtype)
    match mode:
        case "mxfp8":
            elem_dtype = torch.float8_e4m3fn
            choice = MXGemmKernelChoice.CUBLAS
            config = MXFPInferenceConfig(
                activation_dtype=elem_dtype,
                weight_dtype=elem_dtype,
                block_size=32,
                gemm_kernel_choice=choice,
            )
        case "mxfp4":
            elem_dtype = torch.float4_e2m1fn_x2
            choice = MXGemmKernelChoice.CUTLASS
            config = MXFPInferenceConfig(
                activation_dtype=elem_dtype,
                weight_dtype=elem_dtype,
                block_size=32,
                gemm_kernel_choice=choice,
            )
        case "nvfp4":
            config = NVFP4InferenceConfig(use_triton_kernel=use_triton)
        case _:
            raise ValueError(f"Unknown mode {mode}")

    quantize_(model, config)
    model = torch.compile(
        model, fullgraph=True, dynamic=False, mode="max-autotune", backend=backend
    )

    for _ in range(5):
        model(inpt)

    # Use profiler for detailed profiling
    with profiler(f"data/mx_fp_8_{"subclass"}_{mode}", with_stack=True) as p:
        for _ in range(5):
            model(inpt)
            p.step()

    # Use benchmark function for pretty printed runtime
    runtime_us = benchmark_do_bench_in_microseconds(lambda: model(inpt))
    print(f"[bold green]{mode} Runtime:[/bold green] {runtime_us:.2f} μs per iteration")


if __name__ == "__main__":
    CLI(main)

```

> training wouldn't benefit from this, at least without fused quantize+scale_swizzle
Yup this is a proxy for inference where weights are static,  like you said for training since the scales change at every grad step, I dont think this would be much help